dynamically award rsc

+++ src/reputation/models.py
@@ -68,7 +68,7 @@ class Distribution(SoftDeletableModel, PaidStatusModelMixin):
         Hub,
         related_name='reputation_records',
     )
-    amount = models.IntegerField(default=0)
+    amount = models.BigIntegerField(default=0)
     created_date = models.DateTimeField(auto_now_add=True)
     updated_date = models.DateTimeField(auto_now=True)
     distribution_type = models.CharField(
